requirements.txt
# SnapPaste Python Dependencies

# Web 框架
flask>=2.0.0

# 二维码生成
qrcode>=7.0

# 图片处理（内存操作）
Pillow>=9.0.0

# Windows 剪贴板（仅 Windows 平台需要）
pywin32>=300; sys_platform == 'win32'
